+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -982,6 +982,19 @@ schema = Schema((
         schemata="data",
         write_permission="PloneMeeting: Write risky config",
     ),
+    BooleanField(
+        name='removeAnnexesPreviewsOnMeetingClosure',
+        default=defValues.removeAnnexesPreviewsOnMeetingClosure,
+        widget=BooleanField._properties['widget'](
+            description="RemoveAnnexesPreviewsOnMeetingClosure",
+            description_msgid="remove_annexes_previews_on_meeting_closure_descr",
+            label='Removeannexespreviewsonmeetingclosure',
+            label_msgid='PloneMeeting_label_removeAnnexesPreviewsOnMeetingClosure',
+            i18n_domain='PloneMeeting',
+        ),
+        schemata="data",
+        write_permission="PloneMeeting: Write risky config",
+    ),
     TextField(
         name='cssClassesToHide',
         default=defValues.cssClassesToHide,
@@ -6060,6 +6073,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             i = i + 1
             for advice in item.adviceIndex.itervalues():
                 advice['isConfidential'] = adviceConfidentialityDefault
+        logger.info('Done.')
         api.portal.show_message('Done.', request=self.REQUEST)
         return self.REQUEST.RESPONSE.redirect(self.REQUEST['HTTP_REFERER'])