+++ CHANGES.rst
@@ -12,7 +12,7 @@ Changelog
 - Added new type of presence for item attendee (used to ignore an attendee on some items) :
 
   - new meeting optional attribute `non attendee`;
-  - may be used in addition to `present/absent/excused` as even an absent attendee may be set non attendee for a specific item
+  - may be used in addition to `present/absent/excused` as even an absent attendee may be set non attendee for a specific item;
   - changed parameter `patterns` on `print_in_and_out_attendees` to `custom_patterns` to be able to redefine only one single pattern
 - Fixed `AskedAdvicesVocabulary` ram.cache cachekey to avoid same vocabulary used for 2 different MeetingConfigs
   (the `indexAdvisers` term on DashboardCollection was using another MeetingConfig values), moreover made it more robust in case weird context is received
@@ -25,6 +25,8 @@ Changelog
   by default for Meeting related methods.  This way we avoid as much as possible hidden permission problems
 - Exclude SearchableText indexing for IAnnex objects
 - Make sure CKeditor panels are dispayed correctly in popups (adding/editing advice)
+- Added `MeetingConfig.removeAnnexesPreviewsOnMeetingClosure` parameter, when True, annexes previews will be deleted upon meeting closure,
+  added also action on portal_plonemeeting to be able to remove every annexes previews of every items in every closed meetings
 
 4.1.19.2 (2020-03-17)
 ---------------------